# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Code Executor API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
        "/result/{task_id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Task ID",
                        "name": "task_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.ResultResponse"
                        }
                    }
                }
            }
        },
        "/status/{task_id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Task ID",
                        "name": "task_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.StatusResponse"
                        }
                    }
                }
            }
        },
        "/task": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "description": "Task data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Request"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/domain.IdResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.IdResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "domain.Request": {
            "type": "object",
            "properties": {
                "compiler": {
                    "type": "string"
                },
                "program": {
                    "type": "string"
                }
            }
        },
        "domain.ResultResponse": {
            "type": "object",
            "properties": {
                "result": {
                    "type": "string"
                }
            }
        },
        "domain.StatusResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "$ref": "#/definitions/domain.TaskStatus"
                }
            }
        },
        "domain.TaskStatus": {
            "type": "string",
            "enum": [
                "created",
                "in_progress",
                "ready",
                "failed"
            ],
            "x-enum-varnames": [
                "Created",
                "InProgress",
                "Ready",
                "Failed"
            ]
        }
    }
}